+++ src/cuda-host.h
@@ -47,7 +47,7 @@ extern "C"
     CUmodule module;
     CUfunction kernel;
     unsigned char *previous_header;
-    int64_t *timestamp;
+    long long *timestamp;
     double *matrix;
     unsigned char *target;
     char *printf_buffer;       // Device printf buffer
@@ -67,7 +67,7 @@ extern "C"
   void cleanup_cuda_resources(CudaResources *resource);
   void cleanup_all_cuda_gpus(CudaResources *resources, unsigned int device_count);
   cudaError_t run_cuda_hoohash_kernel(CudaResources *resource, unsigned char *previous_header, unsigned char *target, double matrix[64][64],
-                                      int64_t timestamp, uint64_t start_nonce, CudaResult *result);
+                                      long long timestamp, unsigned long long start_nonce, CudaResult *result);
   cudaError_t retrieve_kernel_printf(CudaResources *resource);
 
 #ifdef __cplusplus
